Move genrateRandomMessage code to ExperimentHelper
